# api-docs/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: School Management API
  version: 1.0.0
  description: |
    A comprehensive REST API for managing school operations including classes, students, teachers, subjects, lesson schedules, and attendance tracking.
    
    ## Features
    - **Class Management**: Create and manage school classes with grade levels and letters
    - **Student Management**: Handle student enrollment and information
    - **Teacher Management**: Manage teacher profiles and assignments
    - **Subject Management**: Organize school subjects and curriculum
    - **Lesson Scheduling**: Plan and manage weekly lesson schedules
    - **Lesson Logging**: Track actual lessons that occurred
    - **Attendance & Grading**: Record student attendance and grades
    - **Teacher Assignments**: Link teachers to specific subjects
    
    ## Data Model
    The API follows a relational data model where:
    - Students belong to Classes
    - Teachers are assigned to Subjects via TeacherAssignments
    - LessonSchedules define when subjects are taught to classes
    - LessonLogs record actual lessons that happened
    - StudentLessons track individual student participation, grades, and attendance
servers:
  - url: https://d5ds9jru50sk4al4aiaa.svoluuab.apigw.yandexcloud.net
    description: Local development server
...
paths:
...
  /api/v1/teachers:
    get:
      tags:
        - Teachers
      summary: List all teachers
      operationId: listTeachers
      parameters:
        - in: query
          name: first_name
          schema:
            type: string
          description: Filter by first name (partial match)
        - in: query
          name: last_name
          schema:
            type: string
          description: Filter by last name (partial match)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of teachers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Teacher"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Teachers
      summary: Create a new teacher
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Teacher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/teachers/{id}":
    get:
      tags:
        - Teachers
      summary: Get teacher by ID
      operationId: getTeacher
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Teachers
      summary: Update teacher
      operationId: updateTeacher
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "200":
          description: Teacher updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Teachers
      summary: Delete teacher
      operationId: deleteTeacher
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Teacher deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/subjects:
    get:
      tags:
        - Subjects
      summary: List all subjects
      operationId: listSubjects
      parameters:
        - in: query
          name: subject_name
          schema:
            type: string
          description: Filter by subject name (partial match)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of subjects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subject"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Subjects
      summary: Create a new subject
      operationId: createSubject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/subjects/{id}":
    get:
      tags:
        - Subjects
      summary: Get subject by ID
      operationId: getSubject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Subjects
      summary: Update subject
      operationId: updateSubject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "200":
          description: Subject updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Subjects
      summary: Delete subject
      operationId: deleteSubject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subject deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/teacher-assignments:
    get:
      tags:
        - TeacherAssignments
      summary: List all teacher assignments
      operationId: listTeacherAssignments
      parameters:
        - in: query
          name: teacher_id
          schema:
            type: integer
          description: Filter by teacher ID
        - in: query
          name: subject_id
          schema:
            type: integer
          description: Filter by subject ID
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of teacher assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeacherAssignment"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - TeacherAssignments
      summary: Create a new teacher assignment
      operationId: createTeacherAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherAssignmentInput"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAssignment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/teacher-assignments/{id}":
    get:
      tags:
        - TeacherAssignments
      summary: Get teacher assignment by ID
      operationId: getTeacherAssignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAssignment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - TeacherAssignments
      summary: Update teacher assignment
      operationId: updateTeacherAssignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherAssignmentInput"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAssignment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - TeacherAssignments
      summary: Delete teacher assignment
      operationId: deleteTeacherAssignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Assignment deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the teacher
        first_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Teacher's first name
        last_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Teacher's last name
        patronymic:
          type: string
          maxLength: 50
          description: Teacher's patronymic (optional)
      required:
        - first_name
        - last_name
      example:
        id: 1
        first_name: "Мария"
        last_name: "Петрова"
        patronymic: "Сергеевна"
    Subject:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the subject
        subject_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the subject
      required:
        - subject_name
      example:
        id: 1
        subject_name: "Математика"
    TeacherAssignment:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the assignment
        teacher_id:
          type: integer
          description: ID of the assigned teacher
        subject_id:
          type: integer
          description: ID of the assigned subject
      required:
        - teacher_id
        - subject_id
      example:
        id: 1
        teacher_id: 1
        subject_id: 1
...
    TeacherInput:
      type: object
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Teacher's first name
        last_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Teacher's last name
        patronymic:
          type: string
          maxLength: 50
          description: Teacher's patronymic (optional)
      required:
        - first_name
        - last_name
    SubjectInput:
      type: object
      properties:
        subject_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the subject
      required:
        - subject_name
    TeacherAssignmentInput:
      type: object
      properties:
        teacher_id:
          type: integer
          description: ID of the assigned teacher
        subject_id:
          type: integer
          description: ID of the assigned subject
      required:
        - teacher_id
        - subject_id
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
      required:
        - page
        - limit
        - total
        - pages
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
      required:
        - error
        - message

  # Reusable Response Definitions
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "BadRequest"
            message: "Invalid input parameters"
            details:
              field: "grade"
              issue: "Must be between 1 and 12"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "NotFound"
            message: "Resource not found"
    UnprocessableEntity:
      description: Unprocessable Entity - Business logic validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "UnprocessableEntity"
            message: "Business logic validation failed"
            details:
              field: "teacher_id"
              issue: "Teacher is already assigned to this subject"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"